frontend/requirements.txt
streamlit
google-api-python-client
requests
pandas
langchain
google-generativeai
python-dotenv